Moving generic to sqlite package
