+++ pkg/kine/drivers/sqlite/metrics.go
@@ -1,4 +1,4 @@
-package generic
+package sqlite
 
 import (
 	"time"
